Support writing to a global swift index store
"-global-index-store-path" doesn't change where the index store is output to on its own, it's only used if at that path a file named "rules_swift_global_index_enabled" exists, which needs to be managed by an external process.

Always passing this flag allows caching of the results if they are output to the non-global location. This allows for incremental compilation to use the "rules_swift_global_index_enabled" file and merge cached a fresh results (outside of Bazel).